Improve overloads
