+++ packages/hardhat-viem/test/fixture-projects/type-generation/snapshots/contracts/A.sol/B.ts
@@ -32,31 +32,6 @@ export interface B$Type {
   "deployedLinkReferences": {}
 }
 
-export const B: B$Type = {
-  "_format": "hh-sol-artifact-1",
-  "contractName": "B",
-  "sourceName": "contracts/A.sol",
-  "abi": [
-    {
-      "inputs": [],
-      "name": "getB",
-      "outputs": [
-        {
-          "internalType": "uint256",
-          "name": "",
-          "type": "uint256"
-        }
-      ],
-      "stateMutability": "pure",
-      "type": "function"
-    }
-  ],
-  "bytecode": "0x608060405234801561001057600080fd5b5060b68061001f6000396000f3fe6080604052348015600f57600080fd5b506004361060285760003560e01c8063a1c5191514602d575b600080fd5b60336047565b604051603e91906067565b60405180910390f35b60006002905090565b6000819050919050565b6061816050565b82525050565b6000602082019050607a6000830184605a565b9291505056fea264697066735822122061786bba65c8a92be0aac5ece7c797076fe8ac8905010459067ba24d4ecebd9b64736f6c63430008130033",
-  "deployedBytecode": "0x6080604052348015600f57600080fd5b506004361060285760003560e01c8063a1c5191514602d575b600080fd5b60336047565b604051603e91906067565b60405180910390f35b60006002905090565b6000819050919050565b6061816050565b82525050565b6000602082019050607a6000830184605a565b9291505056fea264697066735822122061786bba65c8a92be0aac5ece7c797076fe8ac8905010459067ba24d4ecebd9b64736f6c63430008130033",
-  "linkReferences": {},
-  "deployedLinkReferences": {}
-};
-
 declare module "@nomicfoundation/hardhat-viem/types" {
   export function deployContract(
     contractName: "contracts/A.sol:B",
