+++ packages/hardhat-viem/test/fixture-projects/type-generation/snapshots/contracts/C.sol/C.ts
@@ -32,31 +32,6 @@ export interface C$Type {
   "deployedLinkReferences": {}
 }
 
-export const C: C$Type = {
-  "_format": "hh-sol-artifact-1",
-  "contractName": "C",
-  "sourceName": "contracts/C.sol",
-  "abi": [
-    {
-      "inputs": [],
-      "name": "getC",
-      "outputs": [
-        {
-          "internalType": "uint256",
-          "name": "",
-          "type": "uint256"
-        }
-      ],
-      "stateMutability": "pure",
-      "type": "function"
-    }
-  ],
-  "bytecode": "0x608060405234801561001057600080fd5b5060b68061001f6000396000f3fe6080604052348015600f57600080fd5b506004361060285760003560e01c8063a2375d1e14602d575b600080fd5b60336047565b604051603e91906067565b60405180910390f35b60006003905090565b6000819050919050565b6061816050565b82525050565b6000602082019050607a6000830184605a565b9291505056fea264697066735822122047c1ea5199833309c2261ac9cf25569e6acf133d1cf8dc638f656e33f555635d64736f6c63430008130033",
-  "deployedBytecode": "0x6080604052348015600f57600080fd5b506004361060285760003560e01c8063a2375d1e14602d575b600080fd5b60336047565b604051603e91906067565b60405180910390f35b60006003905090565b6000819050919050565b6061816050565b82525050565b6000602082019050607a6000830184605a565b9291505056fea264697066735822122047c1ea5199833309c2261ac9cf25569e6acf133d1cf8dc638f656e33f555635d64736f6c63430008130033",
-  "linkReferences": {},
-  "deployedLinkReferences": {}
-};
-
 declare module "@nomicfoundation/hardhat-viem/types" {
   export function deployContract(
     contractName: "C",
